suggested updates

+++ src/AssessmentTracker/Models/AssessmentQuestion.cs
@@ -6,7 +6,7 @@ namespace AssessmentTracker.Models
 
 		public int Id { get; set; }
 
-		public int? Score { get; set; }
+		public float? Score { get; set; }
 
 		public string Comments { get; set; }
 	}
